Moving the cursor and type helpers into CursorWriter
